qa/221: extract MessageBubble module; rewrite memoization test with render counter and proven red step
- Extract MessageBubble, BubbleProps, LocalImageMessage, isLocalImageMessage, REACTIONS
  from ConversationScreen.tsx into a standalone MessageBubble.tsx so jest.mock can
  intercept the module boundary
- jest.mock('../MessageBubble') exposes __getRenderCount / __resetRenderCount via a
  closure-based counter on MockBubble
- Memoization test injects one message synchronously via DeviceContext.Provider
  (innermost provider wins over AllProviders) — no async timing, no waitFor
- Red step proven: test fails with renderCount > 0 when React.memo(MessageRow) is
  absent; passes when present
- Lint: 0 errors (1 pre-existing tzTick warning); tsc: clean; tests: 376/384 pass,
  8 pre-existing failures, 0 new regressions
